@@ -2,94 +2,65 @@
 using System;
 using System.Collections.Generic;
 
-namespace RestApp.Controllers
+namespace RestApp.Controllers;
+
+public record TimeIntervals(List<List<string>> Time, int ExtraHours, List<string> WorkingHours);
+public record RemainingClockTime(string HorasJornada, string HoraFin, string HorasTrabajadas, string HorasRestantes, bool Overtime);
+public record Response(string UtcNow, List<RemainingClockTime> RemainingClockTime);
+
+[ApiController]
+[Route("time")]
+public class TimeController : ControllerBase
 {
-    public class TimeIntervals
-    {
-        public List<List<string>> Time { get; set; }
-        public int ExtraHours { get; set; } = 0;
-    }
-    public class RemainingClockTime
-    {
-        public string HorasJornada { get; set; }
-        public string HoraFin { get; set; }
-        public string HorasTrabajadas { get; set; }
-        public string HorasRestantes { get; set; }
-        public bool Overtime { get; set; }
-    }
+    private static readonly string _formatTimeSpan = "hh\\:mm";
+    private static readonly string _formatDateTime = "H:mm";
 
-    public class Response
+    [HttpPost]
+    public Response Post([FromBody] TimeIntervals intervals)
     {
-        public string UtcNow { get; set; }
-        public List<RemainingClockTime> RemainingClockTime { get; set; }
-    }
-
-    [ApiController]
-    [Route("time")]
-    public class TimeController : ControllerBase
-    {
-        private static readonly string _formatTimeSpan = "hh\\:mm";
-        private static readonly string _formatDateTime = "H:mm";
-        private static readonly List<KeyValuePair<string, TimeSpan>> _timeSpans = new List<KeyValuePair<string, TimeSpan>>
+        if (intervals == null || intervals.Time == null || intervals.Time.Count == 0)
         {
-            new KeyValuePair<string, TimeSpan>(
-                "08:30",
-                TimeSpan.ParseExact("08:30", _formatTimeSpan, null)
-            ),
-            new KeyValuePair<string, TimeSpan>(
-                "08:00",
-                TimeSpan.ParseExact("08:00", _formatTimeSpan, null)
-            ),
-            new KeyValuePair<string, TimeSpan>(
-                "06:00",
-                TimeSpan.ParseExact("06:00", _formatTimeSpan, null)
-            ),
-        };
-
-        [HttpPost]
-        public Response Post([FromBody] TimeIntervals intervals)
-        {
-            TimeSpan total = new TimeSpan();
-
-            int extraHours = intervals.ExtraHours < 0 ? 0 : intervals.ExtraHours;
-            DateTime now = DateTime.UtcNow.AddHours(extraHours);
-
-            foreach (List<string> interval in intervals.Time)
-            {
-                if (interval.Count > 0)
-                {
-                    TimeSpan left = TimeSpan.ParseExact(interval[0], _formatTimeSpan, null);
-                    TimeSpan right = interval.Count > 1
-                        ? TimeSpan.ParseExact(interval[1], _formatTimeSpan, null)
-                        : TimeSpan.FromTicks(now.Ticks);
-
-                    total += right - left;
-                }
-            }
-
-            TimeSpan _total = new TimeSpan(total.Hours, total.Minutes, total.Seconds);
-            List<RemainingClockTime> remainingClockTimes = new List<RemainingClockTime>();
-
-            foreach (KeyValuePair<string, TimeSpan> timeSpan in _timeSpans)
-            {
-                TimeSpan remaining = timeSpan.Value - _total;
-                bool isOvertime = _total >= timeSpan.Value;
-                string remainingSymbol = isOvertime ? "-" : string.Empty;
-                remainingClockTimes.Add(new RemainingClockTime()
-                {
-                    HorasJornada = timeSpan.Key,
-                    HoraFin = (now + remaining).ToString(_formatDateTime),
-                    HorasTrabajadas = _total.ToString(_formatTimeSpan),
-                    HorasRestantes = $"{remainingSymbol}{remaining.ToString(_formatTimeSpan)}",
-                    Overtime = isOvertime
-                });
-            }
-
-            return new Response()
-            {
-                UtcNow = now.ToString(_formatDateTime),
-                RemainingClockTime = remainingClockTimes
-            };
+            return new Response(string.Empty, new());
         }
+
+        TimeSpan total = new();
+
+        int extraHours = intervals.ExtraHours < 0 ? 0 : intervals.ExtraHours;
+        DateTime now = DateTime.UtcNow.AddHours(extraHours);
+
+        foreach (List<string> interval in intervals.Time)
+        {
+            if (interval.Count > 0)
+            {
+                TimeSpan left = TimeSpan.ParseExact(interval[0], _formatTimeSpan, null);
+                TimeSpan right = interval.Count > 1
+                    ? TimeSpan.ParseExact(interval[1], _formatTimeSpan, null)
+                    : TimeSpan.FromTicks(now.Ticks);
+
+                total += right - left;
+            }
+        }
+
+        TimeSpan _total = new(total.Hours, total.Minutes, total.Seconds);
+        List<RemainingClockTime> remainingClockTimes = new();
+
+        List<string> workingHoursIntervals = intervals.WorkingHours ?? new List<string>() { "08:30", "06:00" };
+
+        foreach (string workingHours in workingHoursIntervals)
+        {
+            TimeSpan workingHourTS = TimeSpan.ParseExact(workingHours, _formatTimeSpan, null);
+            TimeSpan remaining = workingHourTS - _total;
+            bool isOvertime = _total >= workingHourTS;
+            string remainingSymbol = isOvertime ? "-" : string.Empty;
+            remainingClockTimes.Add(new RemainingClockTime(
+                workingHours,
+                (now + remaining).ToString(_formatDateTime),
+                _total.ToString(_formatTimeSpan),
+                $"{remainingSymbol}{remaining.ToString(_formatTimeSpan)}",
+                isOvertime
+            ));
+        }
+
+        return new Response(now.ToString(_formatDateTime), remainingClockTimes);
     }
 }
